Refactor sweeps runner and enhance result handling
- Removed the obsolete sweeps_runner_copy.py file.
- Updated sweeps_runner.py to integrate new Pydantic models for result validation.
- Enhanced the export_results method in FileResultDestination to utilize Pydantic for structured data handling.
- Improved argument validation and configuration management in sweeps_runner.py.
- Added logging for better traceability of vector file handling and execution results.
