--- tests/sweep_framework/sweeps_runner_copy.py
@@ -1,715 +1,0 @@
-import argparse
-import sys
-import os
-import pathlib
-import enlighten
-import importlib
-import datetime as dt
-from multiprocessing import Process, Queue
-from queue import Empty
-from framework.statuses import VectorValidity, TestStatus
-import framework.tt_smi_util as tt_smi_util
-from framework.sweeps_logger import sweeps_logger as logger
-from framework.vector_source import VectorSourceFactory
-from framework.serialize import deserialize
-import subprocess
-from dataclasses import dataclass
-from typing import Optional
-from framework.result_destination import ResultDestinationFactory
-
-PROCESS_TERMINATION_TIMEOUT_SECONDS = 5
-
-
-@dataclass
-class SweepsConfig:
-    """Configuration object for sweeps runner"""
-
-    module_name: Optional[str] = None
-    suite_name: Optional[str] = None
-    vector_source: str = "elastic"
-    file_path: Optional[str] = None
-    vector_id: Optional[str] = None
-    result_destination: str = "elastic"
-    watcher: bool = False
-    measure_perf: bool = False
-    measure_device_perf: bool = False
-    dry_run: bool = False
-    sweeps_tag: Optional[str] = None
-    skip_modules: Optional[str] = None
-    skip_on_timeout: bool = False
-    elastic_connection_string: Optional[str] = None
-    elastic_username: Optional[str] = None
-    elastic_password: Optional[str] = None
-
-
-def create_config_from_args(args) -> SweepsConfig:
-    """Create configuration object from parsed arguments"""
-    config = SweepsConfig(
-        module_name=args.module_name,
-        suite_name=args.suite_name,
-        vector_source=args.vector_source,
-        file_path=args.file_path,
-        vector_id=args.vector_id,
-        result_destination=args.result_dest,
-        watcher=args.watcher,
-        measure_perf=args.perf,
-        measure_device_perf=args.device_perf,
-        dry_run=args.dry_run,
-        sweeps_tag=args.tag,
-        skip_modules=args.skip_modules,
-        skip_on_timeout=args.skip_on_timeout,
-    )
-
-    if config.vector_source == "elastic" or config.result_destination == "elastic":
-        from framework.elastic_config import get_elastic_url
-
-        config.elastic_connection_string = get_elastic_url("corp")
-
-        # Acquire once
-        config.elastic_username = os.getenv("ELASTIC_USERNAME")
-        config.elastic_password = os.getenv("ELASTIC_PASSWORD")
-        if not config.elastic_username or not config.elastic_password:
-            raise ValueError("ELASTIC_USERNAME and ELASTIC_PASSWORD must be set in environment variables")
-
-    return config
-
-
-def validate_arguments(args, parser):
-    # Define validation rules as tuples of (condition, error_message)
-    validation_rules = [
-        # Module name dependencies
-        (args.vector_id and not args.module_name, "Module name is required if vector id is specified."),
-        (args.file_path and not args.module_name, "Module name is required if file path is specified."),
-        (
-            args.vector_source == "file" and not args.module_name,
-            "Module name is required when test vector source is 'file'.",
-        ),
-        # File path constraints
-        (
-            args.file_path and args.vector_source in ["elastic", "vectors_export"],
-            "File path should not be specified when test vector source is 'elastic' or 'vectors_export'.",
-        ),
-    ]
-
-    # Check each validation rule
-    for condition, error_message in validation_rules:
-        if condition:
-            parser.print_help()
-            logger.error(error_message)
-            exit(1)
-
-    # Environment variable validation for elastic database
-    if args.vector_source == "elastic" or args.result_dest == "elastic":
-        elastic_username = os.getenv("ELASTIC_USERNAME")
-        elastic_password = os.getenv("ELASTIC_PASSWORD")
-
-        if not elastic_username or not elastic_password:
-            logger.error("ELASTIC_USERNAME and ELASTIC_PASSWORD must be set in the environment variables.")
-            exit(1)
-
-    # Validate that skip modules is only used when running all modules
-    if args.skip_modules and args.module_name:
-        logger.error("Skip modules is only supported when running all modules.")
-        exit(1)
-
-    logger.info("All argument validations passed successfully.")
-
-
-def get_all_modules():
-    sweeps_path = pathlib.Path(__file__).parent / "sweeps"
-    for file in sorted(sweeps_path.glob("**/*.py")):
-        sweep_name = str(pathlib.Path(file).relative_to(sweeps_path))[:-3].replace("/", ".")
-        yield sweep_name
-
-
-def get_timeout(test_module):
-    try:
-        timeout = test_module.TIMEOUT
-    except:
-        timeout = 30
-    return timeout
-
-
-def sanitize_inputs(test_vectors):
-    info_field_names = ["sweep_name", "suite_name", "vector_id", "input_hash"]
-    header_info = []
-    for vector in test_vectors:
-        header = dict()
-        for field in info_field_names:
-            if field in vector:
-                header[field] = vector.pop(field)
-        if "timestamp" in vector:
-            vector.pop("timestamp")
-        if "tag" in vector:
-            vector.pop("tag")
-        header_info.append(header)
-    return header_info, test_vectors
-
-
-def get_devices(test_module):
-    try:
-        return test_module.mesh_device_fixture()
-    except:
-        return default_device()
-
-
-def get_hostname():
-    return subprocess.check_output(["uname", "-n"]).decode("ascii").strip()
-
-
-def get_username():
-    return os.environ["USER"]
-
-
-def git_hash():
-    try:
-        return subprocess.check_output(["git", "rev-parse", "--short", "HEAD"]).decode("ascii").strip()
-    except Exception as e:
-        return "Couldn't get git hash!"
-
-
-def get_git_author():
-    """Get the git author name"""
-    try:
-        return subprocess.check_output(["git", "config", "user.name"]).decode("ascii").strip()
-    except Exception as e:
-        return "Unknown"
-
-
-def get_git_branch():
-    """Get the current git branch name"""
-    try:
-        return subprocess.check_output(["git", "rev-parse", "--abbrev-ref", "HEAD"]).decode("ascii").strip()
-    except Exception as e:
-        return "Unknown"
-
-
-def get_initiated_by():
-    """Get the user who initiated the run - username for dev, CI pipeline name for CI/CD"""
-    # Check if we're in a CI environment
-    ci_pipeline = os.getenv("GITHUB_WORKFLOW") or os.getenv("CI_PIPELINE_NAME")
-    if ci_pipeline:
-        return ci_pipeline
-    else:
-        return get_username()
-
-
-def run(test_module, input_queue, output_queue, config: SweepsConfig):
-    device_generator = get_devices(test_module)
-    try:
-        device, device_name = next(device_generator)
-        logger.info(f"Opened device configuration, {device_name}.")
-    except AssertionError as e:
-        output_queue.put([False, "DEVICE EXCEPTION: " + str(e), None, None])
-        return
-    try:
-        while True:
-            test_vector = input_queue.get(block=True, timeout=1)
-            # Deserialize a test vector (dictionary) back into original object form
-            if config.result_destination == "postgres" or config.result_destination == "results_export":
-                test_vector = deserialize_vector_for_postgres(test_vector)
-            else:  # DATABASE_BACKEND == "elastic"
-                test_vector = deserialize(test_vector)
-            try:
-                results = test_module.run(**test_vector, device=device)
-                if type(results) == list:
-                    status, message = results[0]
-                    e2e_perf = results[1] / 1000000  # Nanoseconds to milliseconds
-                else:
-                    status, message = results
-                    e2e_perf = None
-            except Exception as e:
-                status, message = False, str(e)
-                e2e_perf = None
-            if config.measure_device_perf:
-                perf_result = gather_single_test_perf(device, status)
-                message = get_updated_message(message, perf_result)
-                output_queue.put([status, message, e2e_perf, perf_result])
-            else:
-                output_queue.put([status, message, e2e_perf, None])
-    except Empty as e:
-        try:
-            # Run teardown in mesh_device_fixture
-            next(device_generator)
-        except StopIteration:
-            logger.info(f"Closed device configuration, {device_name}.")
-
-
-def execute_suite(test_module, test_vectors, pbar_manager, suite_name, module_name, header_info, config: SweepsConfig):
-    results = []
-    input_queue = Queue()
-    output_queue = Queue()
-    p = None
-    timeout = get_timeout(test_module)
-    suite_pbar = pbar_manager.counter(total=len(test_vectors), desc=f"Suite: {suite_name}", leave=False)
-    arch = ttnn.get_arch_name()
-    reset_util = tt_smi_util.ResetUtil(arch)
-
-    if len(test_vectors) > 1 and not config.dry_run:
-        p = Process(target=run, args=(test_module, input_queue, output_queue, config))
-        p.start()
-
-    for i, test_vector in enumerate(test_vectors):
-        vector_id = header_info[i].get("vector_id", "N/A")
-        logger.info(f"Executing test: Module='{module_name}', Suite='{suite_name}', Vector ID='{vector_id}'")
-        if config.dry_run:
-            logger.info(f"Would have executed test for vector {test_vector}")
-            suite_pbar.update()
-            continue
-        result = dict()
-
-        result["start_time_ts"] = dt.datetime.now()
-
-        # Capture the original test vector data BEFORE any modifications
-        original_vector_data = test_vector.copy()
-
-        # Use appropriate deserialization based on database backend
-        validity = deserialize(test_vector["validity"])
-
-        if validity == VectorValidity.INVALID:
-            result["status"] = TestStatus.NOT_RUN
-            result["exception"] = "INVALID VECTOR: " + test_vector["invalid_reason"]
-            result["e2e_perf"] = None
-        else:
-            test_vector.pop("invalid_reason")
-            test_vector.pop("status")
-            test_vector.pop("validity")
-
-            try:
-                if config.measure_perf:
-                    # Run one time before capturing result to deal with compile-time slowdown of perf measurement
-                    input_queue.put(test_vector)
-                    if p is None:
-                        logger.info(
-                            "Executing test (first run, e2e perf is enabled) on parent process (to allow debugger support) because there is only one test vector. Hang detection is disabled."
-                        )
-                        run(test_module, input_queue, output_queue, config)
-                    output_queue.get(block=True, timeout=timeout)
-                input_queue.put(test_vector)
-                if p is None:
-                    logger.info(
-                        "Executing test on parent process (to allow debugger support) because there is only one test vector. Hang detection is disabled."
-                    )
-                    run(test_module, input_queue, output_queue, config)
-                response = output_queue.get(block=True, timeout=timeout)
-                status, message, e2e_perf, device_perf = (
-                    response[0],
-                    response[1],
-                    response[2],
-                    response[3],
-                )
-                if status and config.measure_device_perf and device_perf is None:
-                    result["status"] = TestStatus.FAIL_UNSUPPORTED_DEVICE_PERF
-                    result["message"] = message
-                elif status and config.measure_device_perf:
-                    result["status"] = TestStatus.PASS
-                    result["message"] = message
-                    result["device_perf"] = device_perf
-                elif status:
-                    result["status"] = TestStatus.PASS
-                    result["message"] = message
-                else:
-                    if "DEVICE EXCEPTION" in message:
-                        logger.error(
-                            "DEVICE EXCEPTION: Device could not be initialized. The following assertion was thrown: "
-                            + message,
-                        )
-                        logger.info("Device error detected. The suite will be aborted after this test.")
-                    if "Out of Memory: Not enough space to allocate" in message:
-                        result["status"] = TestStatus.FAIL_L1_OUT_OF_MEM
-                    elif "Watcher" in message:
-                        result["status"] = TestStatus.FAIL_WATCHER
-                    else:
-                        result["status"] = TestStatus.FAIL_ASSERT_EXCEPTION
-                    result["exception"] = message
-                if e2e_perf and config.measure_perf:
-                    result["e2e_perf"] = e2e_perf
-                else:
-                    result["e2e_perf"] = None
-            except Empty as e:
-                if p:
-                    logger.warning(f"TEST TIMED OUT, Killing child process {p.pid} and running tt-smi...")
-                    p.terminate()
-                    p.join(PROCESS_TERMINATION_TIMEOUT_SECONDS)  # Wait for graceful process termination
-                    if p.is_alive():
-                        logger.error(f"Child process {p.pid} did not terminate, killing it.")
-                        p.kill()
-                        p.join()
-                    p = None
-                    reset_util.reset()
-
-                result["status"], result["exception"] = TestStatus.FAIL_CRASH_HANG, "TEST TIMED OUT (CRASH / HANG)"
-                result["e2e_perf"] = None
-                result["original_vector_data"] = original_vector_data
-                result["end_time_ts"] = dt.datetime.now()
-                result["timestamp"] = dt.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
-                result["host"] = get_hostname()
-                result["user"] = get_username()
-                suite_pbar.update()
-                results.append(result)
-
-                # Check if we should skip remaining tests in the suite
-                if config.skip_on_timeout:
-                    # Skip all remaining tests in the suite
-                    logger.info("Skipping remaining tests in suite due to timeout.")
-                    for j in range(i + 1, len(test_vectors)):
-                        remaining_vector = test_vectors[j]
-                        skipped_result = dict()
-                        skipped_result["start_time_ts"] = dt.datetime.now()
-                        skipped_result["original_vector_data"] = remaining_vector.copy()
-                        skipped_result["status"] = TestStatus.NOT_RUN
-                        skipped_result["exception"] = "SKIPPED DUE TO PREVIOUS TIMEOUT"
-                        skipped_result["e2e_perf"] = None
-                        skipped_result["end_time_ts"] = dt.datetime.now()
-                        skipped_result["timestamp"] = dt.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
-                        skipped_result["host"] = get_hostname()
-                        skipped_result["user"] = get_username()
-                        results.append(skipped_result)
-                        suite_pbar.update()
-
-                    # Abort the suite
-                    break
-                else:
-                    logger.info("Continuing with remaining tests in suite despite timeout.")
-                    # Continue to the next test vector without breaking
-
-        # Add the original test vector data to the result
-        result["original_vector_data"] = original_vector_data
-
-        result["end_time_ts"] = dt.datetime.now()
-        result["timestamp"] = dt.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
-        result["host"] = get_hostname()
-        result["user"] = get_username()
-
-        suite_pbar.update()
-        results.append(result)
-
-        # Abort the suite if a fatal device error was encountered
-        if "DEVICE EXCEPTION" in result.get("exception", ""):
-            logger.error("Aborting test suite due to fatal device error.")
-            if p and p.is_alive():
-                p.terminate()
-                p.join()
-            break
-
-    if p is not None:
-        p.join()
-
-    suite_pbar.close()
-    return results
-
-
-def run_sweeps(
-    module_names,
-    config: SweepsConfig,
-    run_contents=None,
-):
-    pbar_manager = enlighten.get_manager()
-
-    # Create vector source based on config
-    if config.vector_source == "elastic":
-        if not all([config.elastic_connection_string, config.sweeps_tag]):
-            raise ValueError("Elastic credentials are required when using elastic vector source")
-
-        # Get credentials from environment
-        elastic_username = config.elastic_username
-        elastic_password = config.elastic_password
-
-        if not all([elastic_username, elastic_password]):
-            raise ValueError("ELASTIC_USERNAME and ELASTIC_PASSWORD must be set in environment variables")
-
-        vector_source = VectorSourceFactory.create_source(
-            config.vector_source,
-            connection_string=config.elastic_connection_string,
-            username=elastic_username,
-            password=elastic_password,
-            tag=config.sweeps_tag,
-        )
-    elif config.vector_source == "file":
-        if not config.file_path:
-            raise ValueError("File path is required when using file vector source")
-        vector_source = VectorSourceFactory.create_source(
-            config.vector_source,
-            file_path=config.file_path,
-        )
-    elif config.vector_source == "vectors_export":
-        vector_source = VectorSourceFactory.create_source(config.vector_source)
-    else:
-        raise ValueError(f"Unknown vector source: {config.vector_source}")
-
-    # Create result destination based on config
-    if config.result_destination == "postgres":
-        result_dest = ResultDestinationFactory.create_destination("postgres")
-    elif config.result_destination == "elastic":
-        if not config.elastic_connection_string:
-            raise ValueError("Elastic connection string is required for elastic result destination")
-
-        elastic_username = config.elastic_username
-        elastic_password = config.elastic_password
-
-        if not all([elastic_username, elastic_password]):
-            raise ValueError("ELASTIC_USERNAME and ELASTIC_PASSWORD must be set for elastic result destination")
-        result_dest = ResultDestinationFactory.create_destination(
-            "elastic",
-            connection_string=config.elastic_connection_string,
-            username=elastic_username,
-            password=elastic_password,
-        )
-    elif config.result_destination == "results_export":
-        result_dest = ResultDestinationFactory.create_destination("results_export")
-    else:
-        raise ValueError(f"Unknown result destination: {config.result_destination}")
-
-    # Initialize run metadata and run record
-    run_id = None
-    final_status = "success"
-
-    if not config.dry_run:
-        run_metadata = {
-            "initiated_by": get_initiated_by(),
-            "host": get_hostname(),
-            "git_author": get_git_author(),
-            "git_branch_name": get_git_branch(),
-            "git_commit_hash": git_hash(),
-            "start_time_ts": dt.datetime.now(),
-            "status": "success",
-            "run_contents": run_contents,
-            "device": ttnn.get_arch_name(),
-        }
-        run_id = result_dest.initialize_run(run_metadata)
-        if run_id:
-            logger.info(f"Initialized run with id: {run_id}")
-
-    # Unified processing regardless of source
-    module_pbar = pbar_manager.counter(total=len(module_names), desc="Modules", leave=False)
-    try:
-        for module_name in module_names:
-            test_module = importlib.import_module("sweeps." + module_name)
-            suites = vector_source.get_available_suites(module_name)
-
-            for suite in suites:
-                suite_start_time = dt.datetime.now()
-
-                vectors = vector_source.load_vectors(module_name, suite)
-                header_info, test_vectors = sanitize_inputs(vectors)
-                results = execute_suite(
-                    test_module, test_vectors, pbar_manager, suite, module_name, header_info, config
-                )
-
-                suite_end_time = dt.datetime.now()
-                logger.info(f"Completed tests for module {module_name}, suite {suite}.")
-
-                # Export results
-                if not config.dry_run and results:
-                    run_context = {
-                        "run_id": run_id,
-                        "test_start_time": suite_start_time,
-                        "test_end_time": suite_end_time,
-                        "git_hash": git_hash(),
-                    }
-                    try:
-                        test_status = result_dest.export_results(header_info, results, run_context)
-                        if test_status == "failure":
-                            final_status = "failure"
-                    except Exception as e:
-                        logger.error(f"Failed to export results for {module_name}, suite {suite}: {e}")
-                        final_status = "failure"
-                        # continue with other suites
-
-            module_pbar.update()
-    except Exception as e:
-        logger.error(f"Error during sweep execution: {e}")
-        final_status = "failure"
-        raise
-    finally:
-        if not config.dry_run:
-            result_dest.finalize_run(run_id, final_status)
-            logger.info(f"Finalized run with status: {final_status}")
-        module_pbar.close()
-
-
-def get_module_names(config: SweepsConfig):
-    """Extract module names based on configuration"""
-    if config.module_name:
-        # Always support comma-separated module names now
-        if "," in config.module_name:
-            module_names = [name.strip() for name in config.module_name.split(",")]
-            logger.info(f"Running multiple modules: {module_names}")
-        else:
-            module_names = [config.module_name]
-            logger.info(f"Running module: {module_names}")
-    else:
-        module_names = list(get_all_modules())
-        logger.info(f"Running all modules.")
-        if config.skip_modules:
-            skip_modules_set = {name.strip() for name in config.skip_modules.split(",")}
-            module_names = [name for name in module_names if name not in skip_modules_set]
-            logger.info(f"But skipping: {', '.join(skip_modules_set)}")
-
-    return module_names
-
-
-def get_run_contents(config: SweepsConfig):
-    """Generate run contents description based on configuration"""
-    if config.module_name or config.suite_name:
-        run_contents_details = []
-        if config.module_name:
-            run_contents_details.append(f"{config.module_name}")
-        if config.suite_name:
-            run_contents_details.append(f"{config.suite_name}")
-        return ", ".join(run_contents_details)
-    else:
-        return "all_sweeps"
-
-
-def enable_watcher():
-    logger.info("Enabling Watcher")
-    os.environ["TT_METAL_WATCHER"] = "120"
-    os.environ["TT_METAL_WATCHER_APPEND"] = "1"
-
-
-def disable_watcher():
-    logger.info("Disabling Watcher")
-    os.environ.pop("TT_METAL_WATCHER")
-    os.environ.pop("TT_METAL_WATCHER_APPEND")
-
-
-def enable_profiler():
-    logger.info("Enabling Device Profiler")
-    os.environ["TT_METAL_DEVICE_PROFILER"] = "1"
-    os.environ["ENABLE_TRACY"] = "1"
-
-
-def disable_profiler():
-    logger.info("Disabling Device Profiler")
-    os.environ.pop("TT_METAL_DEVICE_PROFILER")
-    os.environ.pop("ENABLE_TRACY")
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(
-        prog="Sweep Test Runner",
-        description="Run test vector suites from generated vector database.",
-    )
-    parser.add_argument(
-        "--module-name",
-        required=False,
-        help="Test Module Name(s). When vector source is 'elastic' or 'vectors_export', this can be a comma-separated list (e.g., 'eltwise.unary.relu.relu,matmul.short.matmul').",
-    )
-    parser.add_argument("--suite-name", required=False, help="Suite of Test Vectors to run, or all tests if omitted.")
-
-    parser.add_argument(
-        "--vector-source",
-        required=False,
-        default="elastic",
-        choices=["elastic", "file", "vectors_export"],
-        help="Test vector source. Available presets are ['elastic', 'file', 'vectors_export']",
-    )
-
-    parser.add_argument("--file-path", required=False, help="Read and execute test vectors from a specified file path.")
-
-    parser.add_argument(
-        "--vector-id", required=False, help="Specify vector id with a module name to run an individual test vector."
-    )
-
-    parser.add_argument(
-        "--result-dest",
-        required=False,
-        default="postgres",
-        choices=["elastic", "postgres", "results_export"],
-        help="Specify test result destination. Available presets are ['elastic', 'postgres', 'results_export']",
-    )
-
-    parser.add_argument(
-        "--watcher", action="store_true", required=False, help="Add this flag to run sweeps with watcher enabled."
-    )
-    parser.add_argument(
-        "--perf",
-        action="store_true",
-        required=False,
-        help="Add this flag to measure e2e perf, for op tests with performance markers.",
-    )
-
-    parser.add_argument(
-        "--device-perf",
-        required=False,
-        action="store_true",
-        help="Measure device perf using device profiler. REQUIRES PROFILER BUILD!",
-    )
-
-    parser.add_argument(
-        "--dry-run",
-        action="store_true",
-        required=False,
-        help="Add this flag to perform a dry run.",
-    )
-
-    parser.add_argument(
-        "--tag",
-        required=False,
-        default=os.getenv("USER"),
-        help="Custom tag for the vectors you are running. This is to keep copies seperate from other people's test vectors. By default, this will be your username. You are able to specify a tag when generating tests using the generator.",
-    )
-
-    parser.add_argument(
-        "--skip-modules",
-        required=False,
-        help="Comma-separated list of modules to skip when running all modules.",
-    )
-
-    parser.add_argument(
-        "--skip-on-timeout",
-        action="store_true",
-        required=False,
-        help="Skip remaining tests in suite when a test times out. Default behavior is to not skip.",
-    )
-
-    args = parser.parse_args(sys.argv[1:])
-
-    # Argument validation
-    validate_arguments(args, parser)
-
-    # Create sweeps config object
-    config = create_config_from_args(args)
-
-    # Import Elasticsearch if using elastic database
-    if config.vector_source == "elastic" or config.result_destination == "elastic":
-        from elasticsearch import Elasticsearch, NotFoundError
-        from framework.elastic_config import *
-
-    if config.watcher:
-        enable_watcher()
-
-    if config.measure_device_perf:
-        enable_profiler()
-
-    # Generate run contents description
-    run_contents = get_run_contents(config)
-
-    logger.info(
-        f"Running current sweeps with tag: {config.sweeps_tag} using {config.vector_source} test vector source, outputting to {config.result_destination}."
-    )
-
-    if config.skip_on_timeout:
-        logger.info("Timeout behavior: Skip remaining tests in suite when a test times out.")
-    else:
-        logger.info("Timeout behavior: Continue running remaining tests in suite when a test times out.")
-
-    # Parse modules for running specific tests
-    module_names = get_module_names(config)
-
-    from ttnn import *
-    from framework.serialize import *
-    from framework.device_fixtures import default_device
-
-    run_sweeps(
-        module_names,
-        config=config,
-        run_contents=run_contents,
-    )
-
-    if config.watcher:
-        disable_watcher()
-
-    if config.measure_device_perf:
-        disable_profiler()
